Move extension registry to composables/piniaStores and add test helper
