+++ packages/web-pkg/src/composables/piniaStores/index.ts
@@ -1,0 +1,1 @@
+export * from './extensionRegistry'
