+++ packages/web-test-helpers/src/mocks/useExtensionRegistryMock.ts
@@ -1,0 +1,13 @@
+import { mock } from 'jest-mock-extended'
+import { useExtensionRegistry } from 'web-pkg/src/composables/piniaStores'
+
+export const useExtensionRegistryMock = (
+  options: Partial<ReturnType<typeof useExtensionRegistry>> = {}
+): ReturnType<typeof useExtensionRegistry> => {
+  return {
+    ...mock<ReturnType<typeof useExtensionRegistry>>(),
+    registerExtensions: () => undefined,
+    requestExtensions: () => [],
+    ...options
+  }
+}
